requirements.txt
requests
python-rapidjson
black
pre-commit
